Add / modify unit tests
