+++ tests/unit_tests/schema/get_schema_test.py
@@ -1,7 +1,6 @@
 from unittest.mock import MagicMock
 
 from app.repositories.firebase.schema_firebase_repository import SchemaFirebaseRepository
-from app.services.schema.schema_processor_service import SchemaProcessorService
 
 from tests.test_data import schema_test_data
 
@@ -198,71 +197,3 @@ def test_get_schema_with_missing_guid_error(test_client):
 
     assert response.status_code == 400
     assert response.json()["message"] == "Invalid parameter provided"
-
-
-def test_get_schema_metadata_with_incorrect_key(test_client):
-    """
-    Checks that fastAPI return 400 error with appropriate msg
-    when incorrect key is used to query schema metadata
-    at get_schemas_metadata endpoint
-    """
-    response = test_client.get("/v1/schema_metadata?abc=123")
-
-    assert response.status_code == 400
-    assert response.json()["message"] == "Invalid search provided"
-
-
-def test_get_schema_metadata_with_not_found_error(test_client):
-    """
-    Checks that fastAPI return 404 error with apppropriate msg
-    when schema metadata is not found at get_schemas_metadata
-    endpoint
-    """
-    tmp_storage = SchemaFirebaseRepository.get_schema_metadata_collection
-    SchemaFirebaseRepository.get_schema_metadata_collection = MagicMock()
-    SchemaFirebaseRepository.get_schema_metadata_collection.return_value = {}
-
-    response = test_client.get("/v1/schema_metadata?survey_id=123")
-
-    assert response.status_code == 404
-    assert response.json()["message"] == "No results found"
-
-    SchemaFirebaseRepository.get_schema_metadata_collection = tmp_storage
-
-
-def test_get_survey_id_map_200_response(test_client):
-    """
-    When the list of Survey IDs is fetched successfully, the API must return the correct response with 200 status code
-    """
-    SchemaProcessorService.survey_id_map = MagicMock()
-    SchemaProcessorService.get_survey_id_map.return_value = (
-        schema_test_data.test_survey_id_map
-    )
-    response = test_client.get("/v1/survey_list")
-
-    assert response.status_code == 200
-    assert response.json() == schema_test_data.test_survey_id_map
-
-
-def test_get_survey_id_map_404_response(test_client):
-    """
-    When the list of Survey IDs is empty, the API must return the error response with 404 status code
-    """
-    SchemaProcessorService.get_survey_id_map = MagicMock()
-    SchemaProcessorService.get_survey_id_map.return_value = []
-    response = test_client.get("/v1/survey_list")
-
-    assert response.status_code == 404
-    assert response.json()["message"] == "No Survey IDs found"
-
-
-def test_get_survey_id_map_500_response(test_client_no_server_exception):
-    """
-    If the app encounters a global exception, the API must return the error response with 500 status code
-    """
-    SchemaProcessorService.get_survey_id_map = MagicMock(side_effect=Exception)
-
-    response = test_client_no_server_exception.get("/v1/survey_list")
-
-    assert response.status_code == 500
-    assert response.json()["message"] == "Unable to process request"
